+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [2.85.2] 2022-05-02
+
+- Fix toml2csv error log
+- Deployment tips
+  - Allow deployment with pending Apex Jobs
+  - Update Can not find folder
+
 ## [2.85.1] 2022-04-27
 
 - Enhance sfdx hardis:org:retrieve:sources:retrofit command + JSON schema updates